Moved Dialog, added error checking and notifications
